1) Added staticwebapp.config.json and added a "fallback" route to prevent "page not found" errors. 2) Refactored code in preparation of Authentication Code changes. Up until this point, this Angular App has used MSAL (Microsoft Authentication Library).  MSAL is only used for when your app is *NOT* deployed to Azure. For example, I have been testing my code on localhost. However, once the application is deployed to Azure, you need to use this endpoint (/.auth/login/aad).  It's annoying to have 2 different authentication paths in the code, but that is the way it has to be if the code will be hosted in Azure or on-premise.
The refactoring involved moving an auth-config.ts file to MSAL-auth-config.ts.  This file now contains ONLY configuration settings for MSAL,  msalGuardConfig, and msalInterceptorConfig.

I removed endpoints out of the auth-config.ts file and put them in an endpoints.ts file.  This required my importing {endpoints}
into all the services which use httpclient's get.
