+++ staticwebapp.config.json
@@ -1,5 +1,6 @@
 {
   "navigationFallback": {
-    "rewrite": "/index.html"
+    "rewrite": "/index.html",
+    "exclude": [ "*.{css,scss,js,png,gif,ico,jpg,svg}" ]
   }
 }
